Yet more

+++ src/core/unittest/TicketTest.cpp
@@ -131,7 +131,7 @@ TEST(ResumptionTicketTest, ServerEncDec)
             &EncodedServerTicketLength));
 
     ASSERT_NE(EncodedServerTicket, nullptr);
-    ASSERT_NE(EncodedServerTicketLength, 0);
+    ASSERT_NE((uint16_t)EncodedServerTicketLength, 0);
 
     TEST_QUIC_SUCCEEDED(
         QuicCryptoDecodeServerTicket(
